SDK: Fix useRecordAction
Summary:
I noticed that the scripting block couldn't load when I tried using the latest SDK.
This is because useRecordActionData was calling a newly added method that isn't implemented.

Test Plan:
I tested by adding useRecordActionData() to the hello world example and seeing that it crashed
without this diff and succeeded with it.

Reviewers: somehats, jugglinmike

Reviewed By: somehats

Subscribers: minicat

Differential Revision: https://airtable.phacility.com/D25740