sdk: Defer reference to iface injection mechanism
Summary:
In App testing environments, a "mock" version of AirtableInterface (and
the corresponding global `getAirtableInterfaceAtVersion` function) will
be defined when test authors import a JavaScript module designed for
testing (to be introduced in a subsequent patch). In those situtations,
the global function may not be defined when the "injection" code is
initially evaluated. Defer references to that function until the moment
the interface is needed in order to support testing.

Test Plan:
n/a (this change relaxed a technical restriction, but it is
not expected to influence any existing behvior)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D25823