sdk: Facilitate automated App testing
Summary:
Introduce a minimal API designed to allow App developers to write
automated tests for their code.

Introduce an "internal" mock AirtableInterface as a subclass of a more
generic AirtableInterface which integrates with the Jest testing
framework. This structure will allows the mock AirtableInterface to be
used by third-party App maintainers who are writing tests without Jest.

Test Plan:
n/a (this changeset does not alter production behavior, the
functionality it introduces is verified with automated tests)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D25956