sdk: Test the Todo List App, fix bug, & add infra
Summary:
Use the recently-introduced App testing API to write automated tests for
the "Todo List" example App.

Fix a bug in the Todo List App that was caused by it's use of the
`Record#primaryCellValueAsString` property (deprecated in version 0.0.45
and removed in version 1.0.0) instead of `Record#name`. This correction
is included in this patch instead of in a follow-up patch because the
new tests would be significantly less meaningful without it.

Extend the testing infrastructure with approximated implementations of
Hyperbase functionality to support the new tests.

Test Plan: n/a (this patch introduces automated tests)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D26014