sdk: Support Table creation in test environment
Summary:
Currently, the SDK's public API exposes a method to create Tables from
the Base. However, unlike many other model operations,
`createTableAsync` does not apply this operation on the local data
structure optimistically. It depends on the backend to apply the changes
and to provide appropriate mutations. The backend is absent from the
test environment, so operations which the SDK does not perform
optimistically are not performed at all.

Simulate the backend's response to requests for Table creation so that
App code running in the test environment behaves as though the backend
is present and immediately responds successfully.

Test Plan: n/a (this patch includes automated tests)

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D26588