SDK: Fix view deletion crash
Summary:
When using useLoadable or useWatchable with a ViewQueryResult, deleting the view could cause a crash when calling the unload/unwatch method when unmounting.
This fixes the crash by checking if that view is deleted everywhere we would try to get a viewDataStore based on the view's id in order to unload/unwatch. I think this is cleaner than trying to
handle this case explicitly in useLoadable and useWatchable.

Test Plan:
I updated the chart block to use the latest sdk and tried deleting the chosen view with and without
this patch.

Reviewers: somehats, billylittlefield, minicat

Reviewed By: billylittlefield

Subscribers: jugglinmike

Differential Revision: https://airtable.phacility.com/D26417