sdk: Support simulation of Cursor data in test env
Summary:
Extend the App testing interface to allow App developers to simulate
changes to the active table and changes to the selected records.
Demonstrate the usage of the new testing API through a new test suite
for the "Update Records" example App.

Test Plan:
n/a (these changes are verified by new and existing automated
tests)

Reviewers: somehats, benrazon, minicat

Reviewed By: somehats

Subscribers: minicat

Differential Revision: https://airtable.phacility.com/D26426