cli: Support testing Apps outside of App frame
Summary:
Test automation technology like WebDriver and Cypress expect to access
the application under test by navigating a web browser to a URL. Extend
the CLI to meet this expectation via a new flag for the `block run`
command named `--test`.

Test Plan:
execute `block run` in a valid App directory *without* the
new flag and verify that (1) the local development server responds to
GET requests for its index with an inert HTML document, and (2) the App
can be successfully loaded on airtable.com. Execute `block run` in the
Todo List App's directory *with* the new flag and verify that the local
development server hosts a standalone version of the App which
faithfully simulates behavior on airtabe.com.

Reviewers: somehats, benrazon, mzgoddard

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D26561