sdk: Implement BaseProvider to render outside Apps
Summary:
Airtable's Org Chart App attempts to render some of the SDK's React
Components outside of the React tree created by `initializeBlock`.
Recent changes to the SDK [1] subverted this use case. Introduce a React
Provider Component which allows consumers to render in this situation by
providing a reference to an instance of the Base model.

[1] 7fcbfae24378fcc83572f5c01610fd28a0d761d6

Test Plan: n/a (automated by the included tests)

Reviewers: somehats, benrazon

Reviewed By: somehats, benrazon

Differential Revision: https://airtable.phacility.com/D26557