sdk: Deprecate bindings which resist testing
Summary:
For a given App, there is only one instance of the Cursor model and of
the Session model. The SDK exposes these instances via dedicated
bindings in the package's main entry point. Although this pattern may be
acceptable for writing functional applications, it complicates the
process of testing the same code (multiple distinct Base instances may
be created in testing environments).

Extend the `Base` class with instances properties which reference these
instances so that authors who wish to write automated tests for their
Apps can structure their code to be more testable. Update the code
examples to demonstrate this new pattern, and document the old pattern
as "deprecated" in the project's change log.

Test Plan:
n/a (this patch introduces a new feature without modifying
existing functionality, and it includes tests for the behavior it
introduces)

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D26162