CustomViews inside blocks-SDK
Summary:
This adds a new initializeView() entrypoint to the sdk. Blocks can use both the initializeBlock() and initializeView() entrypoints, and is in-fact what the Gantt app does today.

This is mostly untouched from @somehats 's work from hackday 😁 , so all credit goes to Alex 🎉

Not intended to be public yet. So perhaps shouldn't land it.

Depends on D26357

Test Plan: No testing needed

Reviewers: #team-platform-ecosystem, somehats, fredz, billylittlefield

Reviewed By: #team-platform-ecosystem, somehats, fredz, billylittlefield

Subscribers: billylittlefield, benrazon, fredz, somehats

Differential Revision: https://airtable.phacility.com/D26385