Cli: Add logging for remote feature
Summary:
Updates the cli to include some information about the default remote.json file when
releasing a block so that we can gather metrics about how it is being used.

Test Plan:
I tested manually by using a local build and confirming that block release works and
outputs the right metrics. Note that we must not release this until after the corresponding
hyperbase commit is in production.

Reviewers: billylittlefield, minicat

Reviewed By: billylittlefield, minicat

Differential Revision: https://airtable.phacility.com/D26777