sdk: Refactor to support testing `expandRecord`
Summary:
Extend the App testing library and refactor the SDK internals to allow
automated App tests to verify when and how Apps request the expansion of
records.

Test Plan:
For an existing App that is designed to expand records (e.g.
the Todo List example App), verify that the App's build process succeeds
and that it continues to expands records as intended.

Reviewers: somehats, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D26166