sdk: Support early Record creation in test env
Summary:
Once a Table is loaded, the test environment implicitly supports
creating Records in that table (relying on the SDK's ability to
optimistically apply such updates). However, the same functionality is
not available prior to Table loading. Implement support for this
operation so that test authors can consistently create Records
regardless of the state of the App under test.

Test Plan: n/a (this patch includes an automated test)

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D26898