SDK: improve types for colorUtils
Summary:
I noticed that when working with colorUtils in typescript the inferred return type of
certain calls included null even when called with a literal value that should never
return null. By changing the order that the interface defines methods, the non-null
method should match supported constants first and only fall back to including null
in other cases.

Test Plan:
I tested by running a block with the local sdk and then editing calls to getHexForColor
and confirming that I no longer needed to include ! when using a supported constant but
did need to when using an unsupported constant or a string variable

Reviewers: somehats, KateRudolph-at

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D27310