cli-next: stop responses on task channel after teardownAsync
Summary:
- add a close method implementing a Closable interface to the PipeEndChannel
- add a AsyncIterable implmentation to ResponseChannelImplementation
- use the AsyncIterable implementation in task_host.ts

The child process in the conditions of the release command continues to run
after it finishes bundling and is instructed to teardown. The channels that
listen for messages for the remote process instruct node to continue its event
loop. This change removes those listeners after teardown is responded to.

Implementing the ResponseChannelImplementation AsyncIterable interface,
task_host observes requests from the producer process. After a teardown request
is handled by the local the PipeEndChannel is closed, removing listeners from
the node process object keeping the node vent loop alive.

A follow up step may be to after a short delay, in case the process is
misbehaving and is still running, forcefully stop it.

Test Plan:
1. install deps
    - yarn
2. execute unit tests
    - yarn test
3. run command
    - yarn pack:typescript
    - yarn link
    - cd ..
    - git clone https://github.com/Airtable/apps-hello-world-typescript.git
    - cd apps-hello-world-typescript
    - write installed block id into .block/remote.json
    - yarn link @airtable/blocks-cli-next
    - $(npm bin)/block release
    - the process completes without error returning to the shell

Reviewers: somehats, jugglinmike, benrazon

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D27767