sdk: Reduce testing API's reliance on internals
Summary:
The testing API is expected to be relocated to a standalone package.
This patch reduces the amount of internals that will need to be exposed
to enable such a separate.

Test Plan:
n/a (this change is not behavioral, and continued operation
is verified by the existing automated tests)

Reviewers: somehats, benrazon

Reviewed By: somehats, benrazon

Differential Revision: https://airtable.phacility.com/D27805