Update development version of TypeScript
Summary:
The Yarn package manager automatically updates the TypeScript module to
the most recent release that satisfies all dependent packages. Two
"patch" releases of TypeScript were recently published, causing many
uses of the Yarn command-line application to modify to the tool's "lock"
file.

Apply the change to that file in a dedicated commit so that it can be
verified independently from other changes.

Test Plan: n/a

Reviewers: somehats, benrazon, mzgoddard

Reviewed By: benrazon, mzgoddard

Differential Revision: https://airtable.phacility.com/D28035