cli-next: add __runFrame routes
Summary:
- string interpolate block base url in entry point
- add dev server run frame routes
- signal build status to run frame routes from bundler

The run command can be currently unit tested but not used in the airtable app
frame container. This change adds some needed url routes the frame depends on:
`/ping`, `/ping.gif`, `/bundle.js`, `/bundleStatus`.

To support `/bundleStatus`, a method has been added to RunTaskConsumer that the
bundler implements to provide that information.

`/poll_script.js` and `/poll` are not implemented in this change.

This change does not yet support livereload. Currently, by default, webpack's
livereload is active and trying to connect, but the configuration needs a minor
update that will be made in a follow up.

In development mode the entry sets the base uri. The current cli uses envify to
do this. We can use webpack's environment configuration to do the same. But if
we only need it in the entry I think it may be simplier to include it in the
entry point template interpolation.

Test Plan:
1. install deps
  - yarn

2. execute unit tests
  - yarn test

3. run command
  - yarn pack:typescript
  - yarn link
  - cd ..
  - git clone https://github.com/Airtable/apps-hello-world-typescript.git
  - cd apps-hello-world-typescript
  - write installed block id into .block/remote.json
  - yarn link @airtable/blocks-cli-next
  - $(npm bin)/block run
  - open app in airtable
  - put the app in edit mode with https://localhost:9000
  - Hello world :rocket: will be visible

Reviewers: somehats, jugglinmike, benrazon, odomojuli

Reviewed By: somehats, benrazon

Differential Revision: https://airtable.phacility.com/D27776