Update documentation: allow creating linked record fields
Summary:
This adds documentation to the SDK based on the changes in D27850
Since the changes required are hyperbase only, there's no actual functionality change to the SDK and these changes will be available after the relevant hyperbase deploy. However, we maintain the documentation in this repo.

Test Plan: QA can skip

Reviewers: minicat

Reviewed By: minicat

Differential Revision: https://airtable.phacility.com/D28071