sdk: add disabled props for ConfirmationDialog
Summary:
https://airtable.com/tblSbpimM8Spp9txp/viw4S95icG4LLJIt3/rec77XVI43KTEY1Fz?blocks=hide

Requested by a 3p dev on the community forum

Added them for both cancel and confirm as requested. It's a bit weird to add it to cancel, but you can still close the dialog by clicking the X or clicking outside, and it might be useful if you're overriding the default "cancel" text.

Test Plan:
Tested using Storybook

{F263696}

{F263695}

{F263694}

{F263693}

Reviewers: #team-platform-ecosystem, KateRudolph-at

Reviewed By: #team-platform-ecosystem, KateRudolph-at

Subscribers: KateRudolph-at

Differential Revision: https://airtable.phacility.com/D28361