sdk: Improve error message
Summary:
The SDK cannot function in the absence of a valid AirtableInterface
instance. When such an instance cannot be found, it immediately produces
a runtime error. Prior to this commit, the error message was somewhat
misleading because it was written before automated testing was possible.

Re-write the error message to more accurately describe the problem and
to offer guidance on remediation.

Test Plan:
n/a (this change improves developer ergonomics but does not
otherwise effect observable behavior)

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D28379