[Gantt View - Blocks SDK] Enable support for retrieving grouping, and setting groupConfigs
Summary:
(Adding Emma / Ben for perhaps extra SDK insight)
### Summary
This adds the ability to retrieve group information when retrieving a TableOrViewQueryResult (Specifically when it's a view).

The best summary is an example, so here is a block that works with these changes: P211

The other major change introduced that that snippet doesn't cover is the ability to setMetadata for a view, specifically its grouping config (iff it's a Block View). See D28363 for a WIP example of this.

The last change is the ability to access the view's grouping metadata (TODO: well formatted snippet for this).

It mostly mirrors what's been specced out in the [[ https://docs.google.com/document/d/1SSzLe4M3PCqsZg5y-0O7icOweV3Y1q8HRkCkX2QozfU/edit | Grouping API document ]], however setting grouping visibility is missing, which is required (and will be a follow-on diff).

### Open Question
In general this diff has me questioning if this //should// actually extend from RecordQueryResult.

It might be more approachable if it's just it's own concept, some of the pros and cons of it being a RecordQueryResult:

Pros:
 - Adding support in the future to get a specific group result from the backend (this seems.... pretty unlikely to happen anytime soon though)
Cons:
 - Watchable support is a pain to add, and doesn't work currently* - this isn't required, but could be added, to a non-RecordQueryResult model
 - Doesn't really match our other queryResults, it's //not// a direct result from a query, it's more analogous to a record, a model that comes from a query that is executed
    - Because of this, a lot of functions that have to be implemented don't really make sense (like loading data)

One option is to not inherit from RecordQueryResult, give it a name like `GroupedRecords`, and then if we need in the future to add query support `GroupedRecordsQueryResult` can be an alias to `GroupedRecords`, with just a superset of functionality.

* I haven't worked on fully fleshing this out b/c it's fraught with issues (I'd likely disable it) and I'm still not sure if extending from RecordQueryResult is the right call right now

### TODO before landing
Figure out next steps on the above question, and implement tests to cover - this is the main surface area of testing that's missing in this diff, and is the reason tests are failing (due to lack of coverage). In the meantime this section of the diff is also a bit sloppy, but it's enough to show _what_ is needed, and the generally the sloppiness is the sections of code that would be removed if we move away from inheriting RecordQueryResult

We should check is that the public facing docs don't change to include any of this (I attempted to keep everything @hidden).

Depends on D28345
Depends on D28294

Test Plan: Most of the testing around this will come with the follow-on Gantt diff, however - we can test regressions in apps that watch for changes in views that have groups enabled. I can't think of any particular regression in this area, but it's possible?

Reviewers: KateRudolph-at, billylittlefield, fredz, minicat, benrazon, somehats

Reviewed By: KateRudolph-at, billylittlefield

Differential Revision: https://airtable.phacility.com/D28352