sdk: Correct file references in package manifest
Summary:
When npm publishes a package to the registry, it includes the files
named by the `files` property of the package manifest. Prior to this
commit, the SDK package's `files` property did not reference the
`unstable_testing_utils.js` file which is required by the testing
library.

Update the manifest to reference the requisite files and remove
references to non-existent files (these entries are an artifact of a
prior iteration of commit b591a1d03c85ee0005c5fd8ddcc781ad8a96755c).

Test Plan:
1. run the command `npm pack`
2. open the generated archive and verify that it includes the files
   `unstable_testing_utils.js` and `unstable_testing_utils.d.ts`

Reviewers: somehats, benrazon, KateRudolph-at

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D28984