Blocks-Tests: Update blocks sdk dependency
Summary:
We found that the imports in the blocks-testing package still failed using v0.0.2
of the blocks-testing sdk. D28984 fixes this and was released in blocks sdk v1.5.1
so update blocks testing to use that

Test Plan:
I've been having a hard time testing this before releasing, but hopefully this approach
actually exercised everything.
I used npm pack to create a tarball of the blocks-testing package
I updated the batch_update block to use that tarball for blocks-testing and @airtable/blocks@1.5.1
I then ran yarn jest and the tests pass.
Ideally we'd add a test similar to check_typescript_when_installed_in_block.sh to catch this
in jenkins, but I didn't have time to get that working today

Reviewers: somehats, jugglinmike

Reviewed By: jugglinmike

Differential Revision: https://airtable.phacility.com/D29009