cli-next: add error formatting for more human readable errors
Summary:
- add error formatting for more human readable errors
- test VerboseMessage
- cleanup cli action animation in release command

Store error message arguments on the error produced in spawnError. If that error
raises up to the AirtableCommand calling method in the stack, check it for
message arguments and if the first argument appears in the configured Messages
type. If it is format the message to produce a more human digestible message
that can include additional information like urls.

Separating the error formatting from where the error is created we can create
errors with information that is more machine digestible in catch clauses.
Importantly this makes writing tests to check that those errors are thrown
easier. Additionally we can organize all the error formatting into one or a few
files so it is easier to updated the formatted messages without having to for
example search for different files they are in.

Test Plan:
1. install deps
    - yarn
2. execute unit tests
    - yarn test
3. run command
    - yarn pack:typescript
    - yarn link
    - cd ..
    - git clone https://github.com/Airtable/apps-hello-world-typescript.git
    - cd apps-hello-world-typescript
    - write an errorneous installed block id into .block/remote.json
    - yarn link @airtable/blocks-cli-next
    - $(npm bin)/block release
    - the process completes with error

Reviewers: somehats, jugglinmike, benrazon, odomojuli

Reviewed By: somehats, jugglinmike

Differential Revision: https://airtable.phacility.com/D27937