blocks-testing: Explicitly verify compat. with SDK
Summary:
The blocks-testing package documents its compatibility with various
releases of the Blocks SDK via a "peer dependency" in its manifest file.
npm only weakly enforces this restriction (by issuing a warning at the
time of package installation), so consumers may unknowingly install
invalid combinations of the two libraries.

Introduce a runtime compatibility check to ensure that the
blocks-testing library will correctly function with whatever version of
the Blocks SDK is available in the environment (and to produce a
meaningful error message when this is not the case).

This change relies on a previously-untested value exposed by the Blocks
SDK. In order to reduce the risk of future changes which mistakenly
interfere with this integration point, document this dependency with an
inline code comment in the SDK source code and add an automated test to
the SDK test suite.

Test Plan:
The existing automated tests for the Todo example App demonstrate that
this patch does not regress functionality for the supported use cases.
To verify that the improved error message is correctly reported:

1. change the `version` property of the `package.json` file in the SDK
   package from `1.5.1` to `1.4.1`
2. run the script `./scripts/check_integration_with_sdk.sh` from the
   blocks-testing directory

The script should fail with a message describing a version compatibility
problem.

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29033