sdk: Document mutations
Summary:
Prepare for the release of the blocks-testing library by documenting the
Mutation types which will be exposed to developers. Retain the `@hidden`
directives so that this information is not yet emitted in the file
generated by TypeDoc.

Test Plan:
Run the airtable.com Developer site using the `docs.json` file as
modified by this commit and verify that the newly-documented Mutation
types are not visible.

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29238