blocks-docs: Remove unused configuration
Summary:
Prior to this commit, the `sourcefile-url-prefix` value was provided to
the TypeDoc documentation generator, but it is not referenced by the
TypeDoc source code nor by any of the TypeDoc plugins in use by this
project

    $ git log -n1 --format='%H %cI %s'
    77d6543d88493094de9ce6f23fb49739cec61cc8 2021-03-09T18:30:16-05:00 blocks-docs: Enforce linting rules on source code
    $ grep sourcefile-url-prefix . -r
    ./packages/sdk/typedoc.js:    'sourcefile-url-prefix': `https://github.com/airtable/blocks/blob/${releaseTag}/packages/sdk/`,
    ./packages/blocks-testing/typedoc.js:    'sourcefile-url-prefix': `https://github.com/airtable/blocks/blob/${releaseTag}/packages/blocks-testing/`,

This was also the case prior to the recent changes to the documentation
generation process

    $ git checkout 6fe7e30683f772187ab9819d069b7a81a902d5b4
    $ git log -n1 --format='%H %cI %s'
    6fe7e30683f772187ab9819d069b7a81a902d5b4 2021-01-29T11:08:14-08:00 SDK: Add TypedFieldResult to make it easier to type narrow FieldOptions
    $ grep sourcefile-url-prefix . -r
    ./packages/sdk/typedoc.js:    'sourcefile-url-prefix': `https://github.com/airtable/blocks/blob/${releaseTag}/packages/sdk/`,

Remove the configuration to simplify the code and avoid confusion.

Test Plan:
Build each package's documentation with this patch applied and verify
that there is no change to the generated JSON file.

Reviewers: somehats, benrazon, billylittlefield

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29223