blocks-testing: Narrow type of emitted value
Summary:
In addition to the Mutation objects defined by the Blocks SDK, the
`MockAirtableInterface` uses other types of Mutation objects that are
unique to test environments. The interface (and by extension, the
`TestDriver` class) is only intended to emit "mutation" events for the
former type of Mutation. Prior to this commit, however, the signature of
methods like `TestDriver#watch` inaccurately declared that the
test-specific Mutations may be emitted.

The Mutation types defined by the SDK (so-called "authentic" Mutations)
will be publicly documented, but the test-specific Mutations will not.
Some consumers may be confused if the public API references types that
are not available to them.

Narrow the type used to describe the data emitted for "mutation" events.

Test Plan:
n/a (this change is statically verified by the TypeScript
compiler)

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29233