blocks-docs: Generate combined docs for two pkgs
Summary:
Update the API documentation generation process to create a single
JSON-formatted file which includes information about both the `sdk`
package and the `blocks-testing` package. Remove the corresponding
configuration in each of those packages along with the JSON-formatted
documentation files which were previously generated to describe them.

Although the resulting file has more information, it will not change the
documentation available on airtable.com because no members of the
`blocks-testing` have yet been annotated with the `@docsPath` symbol.

Test Plan:
1. Use the `packages/blocks-docs/docs.json` file introduced by this
   commit to run the airtable.com website in a development environment
2. Verify that the site function as expected (that no
   previously-included documentation is omitted, that no new
   documentation is introduced, and that internal and external links
   resolve correctly)

Reviewers: somehats, benrazon, billylittlefield

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29192