blocks-testing: Improve TestDriver documentation
Summary:
Add a code example for every method of the TestDriver class. For
example, include a description of the assumed functionality of the
fictional example App. This is intended to help readers recognize the
distinction between the behavior of the API (which is available for use
in their test environment) and the imaginary behavior under test (which
is not).

Refactor some method signatures so that the generated documentation more
fully describes the API.

Insert explicit links to other parts of the API to improve the
expressiveness of the rendered documentation.

Test Plan: n/a (this change does not meaningfully alter executable code)

Reviewers: somehats, benrazon, mzgoddard

Reviewed By: somehats

Subscribers: odomojuli

Differential Revision: https://airtable.phacility.com/D29303