blocks-testing: Expose MutationTypes
Summary:
Export the SDK's MutationTypes value so that consumers may use it when
working with TestDriver APIs which reference it (e.g. `TestDriver#watch`
and `TestDriver#simulatePermissionCheck`). Add automated tests so that
future maintainers of this library are aware of the contract.

Test Plan: n/a (this change is verified by automated tests)

Reviewers: somehats, benrazon

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29286