cli-next: Silence proxy server on shutdown
Summary:
It is possible for the routine shutdown operation to interrupt requests
while they are being processed by the proxy server. Prior to this
commit, the proxy server would detect this interruption and report an
error, e.g.

    [HPM] Error occurred while trying to proxy request /bundle.js from localhost:9000 to http://localhost:9090 (ECONNREFUSED) (https://nodejs.org/api/errors.html#errors_common_system_errors)

Since this condition does not actually reflect a problem with the
operation of the CLI, the message may mislead some users. Ignore all
messages from the proxy server once the shutdown operation has begun.

Test Plan:
1. Execute the "run" command of the CLI in a valid Block directory on a
   UNIX-like system
2. Wait for the command to print "Bundle updated" to standard output
3. Execute the following command from another terminal window:

    $ curl -k https://localhost:9000/__runFrame/bundle.js -L > /dev/null & kill -s SIGINT $(ps aux | grep -E 'node .*ru[n]' | awk '{print $2}')

4. Verify that the "run" command terminated with a status code of "0"
   and without reporting an error via standard output or standard error

Reviewers: benrazon, mzgoddard, odomojuli, somehats

Reviewed By: somehats

Differential Revision: https://airtable.phacility.com/D29548