blocks-testing: Annotate public API
Summary:
Update the in-line documentation to instruct the documentation
generation process to include information on the public API in the
data structure it creates.

This patch should not be merged immediately because the Blocks Testing
library has not bee released at the time of this writing. The
airtable.com infrastructure should be extended to ignore the information
that this patch exposes; see D29668.

Test Plan: n/a (this change does not modify executable code)

Reviewers: benrazon, billylittlefield, somehats

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D29670