cli-next: Remove unused code
Summary:
This project depends on the `webpack-dev-server` module using the
version specifier `^3.11.0`. At the time of writing, that resolves to
version 3.11.2. At that version, the library's `Server#listen` method
accepts three arguments: a port number, a host name, and a callback
function:

https://github.com/webpack/webpack-dev-server/blob/v3.11.2/lib/Server.js#L774

Prior to this patch, the CLI erroneously provided a callback function
via the second parameter, and the library interpreted this as a host
name. Although this would generally cause the call site to lose
visibility into the completion of the asynchronous "listen" operation,
subsequent seemingly-redundant code is actually tracking the completion
as necessary.

Remove the unused code.

Test Plan:
Run the "run" command in a valid App directory and verify that the CLI
continues to function as expected.

Reviewers: mzgoddard, somehats

Reviewed By: mzgoddard

Differential Revision: https://airtable.phacility.com/D29680