cli: local development over http
Summary:
**Updated description 3/16/21**

Adds a new argument `--http` to the CLI `block run` command, which lets users develop against an HTTP server. This was in response to Chrome 88's temporary removal of the `allow-insecure-localhost` flag, which made it difficult to develop blocks using our default HTTPS strategy. In retrospect, there was a workaround (enable "Temporarily enable M87 flags", then re-enable `allow-insecure-localhost`), but this gives us another alternative.

We chose to not make this the default, because mixed-content is blocked on Safari (other browsers have an exception for localhost), and any hybrid approach is largely complex.

Test Plan:
1. Using this new version of the blocks-cli, verify `block run --http` works (using the new `http` URL)
2. Using this new version of the blocks-cli, verify `block run` works (this should use the existing default `https` behavior)

Reviewers: SeanKeenan-at, somehats, djmally, richsinn

Reviewed By: SeanKeenan-at, richsinn

Subscribers: richsinn, fredz

Differential Revision: https://airtable.phacility.com/D27177