cli-next: Correct & simplify stubbing in test env
Summary:
Prior to this commit, the method in use to simulate process signals was
very brittle. The testing infrastructure captured all event handlers
intended for the global `process` object. Following the beginning of a
given test, it invoked the very first handler to be registered in order
to simulate the signal specified by that test. In this way, it assumed
that the first handler handler bound during test execution would always
be intended for the desired signal. That assumption is not necessarily
true, and it was violated by the introduction of the `clipboardy`
package (which indirectly caused the registration of event handlers for
a number of other signals).

Refactor the internals to tolerate the registration of any number of
process events (and in any order). Use an independent EventEmitter
instance to collect all of the event handlers registered while tests are
running. Simulate process signals by emitting the desired events on that
instance.

Update the application code to avoid a race condition where the signal
could be simulated before the CLI had bound an event listener.

Test Plan:
This commit is not intended to modify the application's functionality,
but because it alters some of the source code, the existing
functionality should be verified to reduce the risk of regression.

1. Clear the system clipboard
2. Execute the command `./bin/run run` from the directory of a valid
   Airable App
3. Verify that the CLI reports that the development server's URL has
   been written to the system clipboard
4. Verify that the reported URL is in fact present in the system
   clipboard
5. Send a SIGINT signal to the CLI process (on many systems, this can be
   accomplished by pressing and holding the "Control" key, then pressing
   the "C" key, then releasing both keys)
6. Verify that the CLI process exits gracefully

Reviewers: benrazon, mzgoddard, somehats

Reviewed By: mzgoddard

Differential Revision: https://airtable.phacility.com/D30296