cli-next: Use security-conscious code generation
Summary:
The Google Chrome web browser will soon enforce more strict rules on web
application behavior, specifically relating to the use of JavaScript's
`eval` built-in function [1]. Change the code generation technique used
in development settings so the output does not make use of that
function.

Add a test for the corresponding "production" build mode. Update the
relevant test fixture file to ensure that the original source is not
completely optimized away.

[1] https://community.airtable.com/t/new-tighter-security-and-network-access-option-for-marketplace-apps/37654

Test Plan:
1. Execute the "run" sub-command from the "Hello World" Example App
2. Download the compiled code from the development server at
   `/bundle.js`. For instance, if the CLI reports:

   > Server listening at https://localhost:9000

   ...then the compiled code can be accessed at the following URL:

       https://localhost:9000/bundle.js

3. Verify that the code does not include any references to the `eval`
   built-in function.

Reviewers: mzgoddard, somehats, benrazon, odomojuli

Reviewed By: mzgoddard

Differential Revision: https://airtable.phacility.com/D29152