cli-next: Enable tests in continuous integration
Summary:
Configure the continuous integration environment to execute all
automated validation tasks: linting, type checking, and unit tests.
Update the project's development scripts to consistently use the `yarn`
command-line interface. Update the application logic to support that
environment:

- disable a security feature designed to mitigate the risk of "shoulder
  surfing" when the environment cannot support it
- tolerate failed attempts to write to the system clipboard

Although both of these modifications alter the behavior of the application in
service of the automated testing environment, neither is expected to regress
functionality for an end-user.

The former modification only applies to scenarios where shoulder surfing cannot
occur (that is: when the standard input stream is not connected to a terminal),
so disabling the feature does not impact security.

The latter modification may apply to end-users who run locked-down systems
where clipboard interaction is limited. Since clipboard interaction is intended
as a convenience, tolerating failure enables such end-users to accomplish the
fundamental task (specifically: the `block run` command).

Test Plan:
1. Run the command `./bin/run set-api-key`
2. Enter the following text: "abcdef"
3. Press the "enter"/"return" key
4. Verify that the text "abcdef" has been replaced with the text "******"
5. Force the process to end by pressing and holding the "control" key and then
   pressing the "C" key
6. Clear the system clipboard
7. Run the command `./bin run run`
8. Verify that the process reports that the development server's URL has been
   copied to the system clipboard
9. Verify that the system clipboard now contains the development server's URL

Reviewers: benrazon, mzgoddard, somehats

Reviewed By: benrazon

Differential Revision: https://airtable.phacility.com/D30324